+++ packages/orq-rc/src/orq_ai_sdk/models/conversationresponse.py
@@ -104,7 +104,7 @@ class ConversationResponse(BaseModel):
     r"""Unix timestamp in milliseconds when the conversation was last modified."""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "conv_01kfmnrkkdt9mwsjs4tqbmcp83"
+        "conv_01kfn442kvjf4z150j2a78aa0n"
     )
     r"""Unique conversation identifier with `conv_` prefix."""