+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -314,7 +314,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KFMNRMHQ37XRKPNZW6ZFDEDN"
+        "01KFN4436ERKR62E2C0NPB6CMF"
     )
     r"""The unique identifier of the data source"""