+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -148,7 +148,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KFMNRMHN68X8285P59NRZJG1"
+        "01KFN4436ADJTM6A4NQXDR152F"
     )
     r"""The unique identifier of the data source"""