+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KFMNRMHT1ZG551J5WT8GX1H0"
+        "01KFN4436JPSWR6FY7QX5X58Z5"
     )
     r"""The unique identifier of the data source"""